+++ tests/test_option_groups.py
@@ -1,8 +1,10 @@
 """Tests for the "option groups" feature/module."""
+from textwrap import dedent
 
 import pytest
 
 import cloup
+from cloup import option
 from tests.util import noop
 
 
@@ -44,3 +46,48 @@ def test_option_group_decorator_raises_if_group_is_passed_to_contained_option():
 def test_option_group_decorator_raises_for_no_options():
     with pytest.raises(ValueError):
         cloup.option_group('grp')
+
+
+@pytest.mark.parametrize(['ctx_value', 'cmd_value', 'should_align'], [
+    pytest.param(True, None, True, id='ctx-yes'),
+    pytest.param(False, None, False, id='ctx-no'),
+    pytest.param(False, True, True, id='none'),
+    pytest.param(True, False, False, id='ctx-yes_cmd-no'),
+    pytest.param(False, True, True, id='ctx-no_cmd-yes'),
+])
+def test_align_option_groups_context_setting(runner, ctx_value, cmd_value, should_align):
+    @cloup.command(
+        context_settings=dict(align_option_groups=ctx_value),
+        align_option_groups=cmd_value,
+    )
+    @cloup.option_group('First group', option('--opt', help='first option'))
+    @cloup.option_group('Second group', option('--much-longer-opt', help='second option'))
+    def cmd(one, much_longer_opt):
+        pass
+
+    result = runner.invoke(cmd, args=('--help',))
+    start = result.output.find('First')
+    if should_align:
+        expected = """
+            First group:
+              --opt TEXT              first option
+
+            Second group:
+              --much-longer-opt TEXT  second option
+
+            Other options:
+              --help                  Show this message and exit."""
+    else:
+        expected = """
+            First group:
+              --opt TEXT  first option
+
+            Second group:
+              --much-longer-opt TEXT  second option
+
+            Other options:
+              --help  Show this message and exit."""
+
+    expected = dedent(expected).strip()
+    end = start + len(expected)
+    assert result.output[start:end] == expected